refactor(knowledge): fix code review issues - extract shared types, split components, fix race conditions
**Code Organization:**
- Extract EditorLanguage type to src/types/editor.ts
- Create src/utils/languageDetection.ts with shared language utilities
- Extract documentUtils.ts with isDocumentEditable helper
- Extract wikiLinkResolver.ts for link resolution logic

**Component Splitting:**
- Split DocumentDetailDialog.tsx (1179 -> 625 lines)
- Create DocumentSummarySection.tsx for summary display
- Create DocumentContentViewer.tsx for content rendering
- Remove duplicate language detection functions

**Bug Fixes:**
- Fix useDocumentDetail state redundancy (fullContent as single source of truth)
- Fix loadMore race condition using offsetRef pattern
- Add AbortController for request cancellation on unmount
- Add yieldToBrowser() to prevent UI blocking in loadAllContent
- Fix JSON preview/edit mode inconsistency by formatting JSON before editing

**Performance:**
- Yield control to browser every 5 chunks during large file loading
- Prevent concurrent loadMore/loadAllContent execution
