+++ frontend/src/__tests__/features/knowledge/document/document-detail-dialog.test.tsx
@@ -59,14 +59,18 @@ jest.mock('@/apis/knowledge-base', () => ({
 jest.mock('@/features/knowledge/document/hooks/useDocumentDetail', () => ({
   useDocumentDetail: () => ({
     detail: {
-      content: 'plain text content',
       content_length: 18,
       truncated: false,
       summary: null,
       chunks: [],
     },
+    fullContent: 'plain text content',
     loading: false,
     error: null,
+    loadingMore: false,
+    hasMoreContent: false,
+    loadMore: jest.fn(),
+    loadAllContent: jest.fn(),
     refresh: jest.fn(),
   }),
 }))
@@ -198,14 +202,18 @@ describe('DocumentDetailDialog wiki-link routing', () => {
     jest.doMock('@/features/knowledge/document/hooks/useDocumentDetail', () => ({
       useDocumentDetail: () => ({
         detail: {
-          content: '[link](../other-kb/doc.md)',
           content_length: 26,
           truncated: false,
           summary: null,
           chunks: [],
         },
+        fullContent: '[link](../other-kb/doc.md)',
         loading: false,
         error: null,
+        loadingMore: false,
+        hasMoreContent: false,
+        loadMore: jest.fn(),
+        loadAllContent: jest.fn(),
         refresh: jest.fn(),
       }),
     }))
