+++ frontend/src/features/knowledge/document/hooks/useDocumentDetail.ts
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-import { useState, useEffect, useCallback } from 'react'
+import { useState, useEffect, useCallback, useRef } from 'react'
 import { knowledgeBaseApi } from '@/apis/knowledge-base'
 import type { DocumentDetailResponse } from '@/types/knowledge'
 import { toast } from 'sonner'
@@ -10,6 +10,17 @@ import { useTranslation } from '@/hooks/useTranslation'
 
 // Maximum characters per request (matches backend MAX_DOCUMENT_READ_LIMIT)
 const MAX_CHARS_PER_REQUEST = 100000
+
+// Maximum number of chunks to process before yielding to prevent UI blocking
+const CHUNKS_BEFORE_YIELD = 5
+
+/**
+ * Yield control back to the browser to prevent UI blocking during large file loading.
+ * Uses setTimeout to allow the browser to process pending rendering and user input.
+ */
+function yieldToBrowser(): Promise<void> {
+  return new Promise(resolve => setTimeout(resolve, 0))
+}
 
 interface UseDocumentDetailOptions {
   kbId: number
@@ -19,9 +30,15 @@ interface UseDocumentDetailOptions {
   enabled?: boolean
 }
 
+// Omit content from detail to prevent confusion - use fullContent instead
+type DocumentDetailWithoutContent = Omit<DocumentDetailResponse, 'content'>
+
 interface UseDocumentDetailReturn {
-  /** Document detail data */
-  detail: DocumentDetailResponse | null
+  /**
+   * Document detail data (metadata, summary).
+   * NOTE: Does NOT include content. Use fullContent for document content.
+   */
+  detail: DocumentDetailWithoutContent | null
   /** Whether initial loading is in progress */
   loading: boolean
   /** Error message if loading failed */
@@ -34,7 +51,7 @@ interface UseDocumentDetailReturn {
   currentOffset: number
   /** Whether more content is available to load */
   hasMoreContent: boolean
-  /** Full accumulated content from all loads */
+  /** Full accumulated content from all loads - SINGLE SOURCE OF TRUTH for content */
   fullContent: string
   /** Refresh document detail */
   refresh: () => Promise<void>
@@ -51,6 +68,11 @@ interface UseDocumentDetailReturn {
  *
  * For large documents (>100k chars), content is loaded in chunks.
  * The hook accumulates content and provides loadMore functionality.
+ *
+ * Features:
+ * - Request cancellation via AbortController when component unmounts
+ * - Race condition prevention using refs for offset tracking
+ * - Single source of truth for content (fullContent)
  */
 export function useDocumentDetail({
   kbId,
@@ -62,7 +84,7 @@ export function useDocumentDetail({
   const { t } = useTranslation('knowledge')
 
   // Main detail state (contains summary and metadata)
-  const [detail, setDetail] = useState<DocumentDetailResponse | null>(null)
+  const [detail, setDetail] = useState<DocumentDetailWithoutContent | null>(null)
 
   // Loading states
   const [loading, setLoading] = useState(true)
@@ -75,10 +97,37 @@ export function useDocumentDetail({
   const [hasMoreContent, setHasMoreContent] = useState(false)
   const [fullContent, setFullContent] = useState('')
 
+  // Use ref to track offset for preventing race conditions
+  const offsetRef = useRef(0)
+  // Use ref for abort controller to cancel requests on unmount
+  const abortControllerRef = useRef<AbortController | null>(null)
+
+  // Update ref when state changes
+  useEffect(() => {
+    offsetRef.current = currentOffset
+  }, [currentOffset])
+
+  // Cleanup abort controller on unmount
+  useEffect(() => {
+    return () => {
+      if (abortControllerRef.current) {
+        abortControllerRef.current.abort()
+        abortControllerRef.current = null
+      }
+    }
+  }, [])
+
   /**
    * Fetch document detail (initial load or full refresh)
    */
   const fetchDetail = useCallback(async () => {
+    // Cancel any pending request
+    if (abortControllerRef.current) {
+      abortControllerRef.current.abort()
+    }
+    // Create new abort controller for this request
+    abortControllerRef.current = new AbortController()
+
     try {
       setLoading(true)
       setError(null)
@@ -90,24 +139,33 @@ export function useDocumentDetail({
         limit: MAX_CHARS_PER_REQUEST,
       })
 
-      setDetail(response)
+      // Extract content from response and store separately
+      // Content is excluded from detail to make fullContent the single source of truth
+      const { content: initialContent, ...detailWithoutContent } = response
+      setDetail(detailWithoutContent)
 
       // Reset pagination state
-      if (response.content !== undefined) {
-        setFullContent(response.content || '')
-        setCurrentOffset(response.content?.length || 0)
+      if (initialContent !== undefined) {
+        const content = initialContent || ''
+        setFullContent(content)
+        setCurrentOffset(content.length)
+        offsetRef.current = content.length
 
         // Determine if more content is available
         // truncated=true means there's more content beyond what was returned
         const contentLength = response.content_length || 0
-        const returnedLength = response.content?.length || 0
-        setHasMoreContent(response.truncated === true || returnedLength < contentLength)
+        setHasMoreContent(response.truncated === true || content.length < contentLength)
       } else {
         setFullContent('')
         setCurrentOffset(0)
+        offsetRef.current = 0
         setHasMoreContent(false)
       }
     } catch (err) {
+      // Don't show error for aborted requests
+      if (err instanceof Error && err.name === 'AbortError') {
+        return
+      }
       const errorMessage = err instanceof Error ? err.message : 'Failed to load document detail'
       setError(errorMessage)
       toast.error(t('document.detail.content.error'))
@@ -118,9 +176,13 @@ export function useDocumentDetail({
 
   /**
    * Load more content (pagination)
+   * Uses ref to prevent race conditions with stale closure values
    */
   const loadMore = useCallback(async () => {
     if (loadingMore || !hasMoreContent) return
+
+    // Use ref to get latest offset value
+    const offset = offsetRef.current
 
     try {
       setLoadingMore(true)
@@ -128,7 +190,7 @@ export function useDocumentDetail({
       const response = await knowledgeBaseApi.getDocumentDetail(kbId, docId, {
         includeContent: true,
         includeSummary: false, // Don't need summary on pagination
-        offset: currentOffset,
+        offset: offset,
         limit: MAX_CHARS_PER_REQUEST,
       })
 
@@ -136,16 +198,18 @@ export function useDocumentDetail({
         // Append new content to existing content
         setFullContent(prev => prev + response.content)
 
-        // Update offset
-        const newOffset = currentOffset + response.content.length
+        // Update offset using ref pattern to avoid stale closure
+        const newOffset = offset + response.content.length
         setCurrentOffset(newOffset)
+        offsetRef.current = newOffset
 
         // Check if there's more content
         const totalLength = response.content_length || 0
-        setHasMoreContent(newOffset < totalLength)
+        const stillHasMore = newOffset < totalLength
+        setHasMoreContent(stillHasMore)
 
-        // Update detail content for consistency
-        setDetail(prev => (prev ? { ...prev, truncated: newOffset < totalLength } : null))
+        // Update detail truncated flag for consistency
+        setDetail(prev => (prev ? { ...prev, truncated: stillHasMore } : null))
       }
     } catch (err) {
       const errorMessage = err instanceof Error ? err.message : 'Failed to load more content'
@@ -153,7 +217,7 @@ export function useDocumentDetail({
     } finally {
       setLoadingMore(false)
     }
-  }, [kbId, docId, currentOffset, hasMoreContent, loadingMore])
+  }, [kbId, docId, hasMoreContent, loadingMore])
 
   /**
    * Load all remaining content (for editing truncated documents)
@@ -169,11 +233,12 @@ export function useDocumentDetail({
     try {
       setLoadingMore(true)
       // Read state into local variables to avoid stale closure issues
-      let offset = currentOffset
+      let offset = offsetRef.current
       let hasMore: boolean = hasMoreContent
       let accumulatedContent = fullContent
 
       // Keep loading until all content is fetched
+      let chunksProcessed = 0
       while (hasMore) {
         const response = await knowledgeBaseApi.getDocumentDetail(kbId, docId, {
           includeContent: true,
@@ -191,6 +256,12 @@ export function useDocumentDetail({
         } else {
           hasMore = false
         }
+
+        // Yield to browser periodically to prevent UI blocking for very large files
+        chunksProcessed++
+        if (chunksProcessed % CHUNKS_BEFORE_YIELD === 0 && hasMore) {
+          await yieldToBrowser()
+        }
       }
 
       // Update state with all loaded content
@@ -198,7 +269,9 @@ export function useDocumentDetail({
       // so direct state updates are safe here
       setFullContent(accumulatedContent)
       setCurrentOffset(offset)
+      offsetRef.current = offset
       setHasMoreContent(false)
+      // Update truncated flag only (content is not stored in detail)
       setDetail(prev => (prev ? { ...prev, truncated: false } : null))
 
       // Return the final content and state for immediate use by callers
@@ -211,7 +284,7 @@ export function useDocumentDetail({
     } finally {
       setLoadingMore(false)
     }
-  }, [kbId, docId, currentOffset, hasMoreContent, fullContent, loadingMore])
+  }, [kbId, docId, hasMoreContent, fullContent, loadingMore])
 
   /**
    * Refresh document summary
@@ -239,9 +312,16 @@ export function useDocumentDetail({
 
   // Reset state when document changes
   useEffect(() => {
+    // Cancel any pending request
+    if (abortControllerRef.current) {
+      abortControllerRef.current.abort()
+      abortControllerRef.current = null
+    }
+
     setDetail(null)
     setFullContent('')
     setCurrentOffset(0)
+    offsetRef.current = 0
     setHasMoreContent(false)
     // Clear transient flags to avoid stale UI state from previous document
     setLoadingMore(false)
