+++ frontend/src/types/editor.ts
@@ -1,0 +1,26 @@
+// SPDX-FileCopyrightText: 2025 Weibo, Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+/**
+ * Editor language types supported by CodeMirror and WysiwygEditor
+ */
+export type EditorLanguage =
+  | 'markdown'
+  | 'json'
+  | 'javascript'
+  | 'typescript'
+  | 'jsx'
+  | 'tsx'
+  | 'python'
+  | 'html'
+  | 'css'
+  | 'yaml'
+  | 'sql'
+  | 'xml'
+  | 'cpp'
+  | 'c'
+  | 'go'
+  | 'java'
+  | 'rust'
+  | 'text'
